requirements.txt
torch
torchvision
transformers
pillow
opencv-python
huggingface_hub
accelerate
safetensors